Extracting user in separate module
