+++ lib/api/db.ex
@@ -5,6 +5,12 @@ defmodule Api.Db do
   def users do
     execute("select users.name, users.surname from users;", [])
     |> (fn r -> r.rows end).()
+    |> to_structs
+  end
+
+  defp to_structs(users) do
+    users
+    |> Enum.map(fn [name, surname] -> %User{name: name, surname: surname} end)
   end
 
   defp execute(query, params) do
