+++ lib/api/user.ex
@@ -1,0 +1,4 @@
+defmodule User do
+  @derive [Poison.Encoder]
+  defstruct [:name, :surname]
+end
